Multiple changes
Since the lab website will be acting as my template, i have renamed the directory.

Renamed script.js to myScript.js since it was not working with that name

Removed reliance on class180.htm to display additional content, and created separate div tags to allow content display to be toggled.

Linked bootstrap to allow the use of buttons and other functionality.

Removed the redundant background div tag.