Multiple changes
This is the first version of my code which featured all of the core requirements. At this point it featured code which allows the user to both add and remove locomotives, create an entirely new list, or alter one from an existing file. It is also the first version i could recover which features the code to download the edited list as a new JSON file, and also removed the storage() function for greater efficiency.